Refactor redux folders
